NEXUS-3442: Made it a bit generic and cleanup
As per discussion with Igor, there is really no need
to log anything at all, as to filter out malicious clients
or IPs, Nexus logs is not the place to stuff this into, there
is access log, that also contains processing time.

So, logging happens only if exception is NOT due to
client dropping connection, and this is detected recursively
(not only up to 1st level as before), since as it turned out,
in case of TemplateRepresentation, the IOEx broken pipe is
wrapped by Jetty EofEx that is wrapped by VelocityException...

So, this detection now happens at one single place, and is
used to decide to log or not.